requirements.txt
pygame~=2.3.0
customtkinter~=5.1.2
mutagen~=1.46.0
Pillow~=9.5.0
pyglet~=2.0.7